requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker==7.0.0
fastapi==0.110.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
libcst==1.2.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
strawberry-graphql==0.222.0
typer==0.11.0
typing-inspect==0.9.0
typing_extensions==4.10.0
unrar==0.4
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
